requirements.txt
discord.py>=2.4.0,<3.0.0
audioop-lts>=0.2.1; python_version >= "3.13"
